Analyzing customer churn in banking
Project Objective
Customer Churn has become a major issue in banking industry since it leads to the loss in the revenue generated by the bank. In current scenario, acquiring new customers has become hard and expensive since there are many different banking offering a variety of offers. So they say that “The best way to avoid customer churn is to know your customers, and the best way to know your customer is through historical and new customer data”.

The PowerBi reports gives insights into which customers are leaving. The insights have been obtained based on the following questions:

How many total customers are there by country?
Customers and their genders
What is the activity status of a customer?
Is the customer a credit card holder?
What is the churn rate based on the products?
What is the age group of customers churning?
What is the account balance of the customers churning?
WHat is the credit score of the customers churning?
The below reports gives us answers to these questions and this will be very useful in making decisons that would help prevent customers from churning and help focus on acquiring and retaining the right customers.

churn1_new

2

3

4

The dashboard could not be published to the web as the PowerBI only allows dashboards to be shared on paid accounts.